1. Adding Qty to Color Converter. 2. Contact Delete Validation implemented. 3. Contact Add Validation Implemented. 4. Dashboard Layout changed.[Sale Total added.] 5. Loading window added in masters.(Async methods added) 6. Add More Purchase implemented. 7. Reports UI changed. 8. Masters Loading Dialog registered.

+++ ElectronicZone.Wpf/Utility/CommonMethods.cs
@@ -39,13 +39,13 @@ namespace ElectronicZone.Wpf.Utility
             return dataTable.Rows.Cast<System.Data.DataRow>().Take(topRow).CopyToDataTable();
         }
 
-        public static decimal GetSum(DataTable dataTable, string columnName)
+        public static double GetSum(DataTable dataTable, string columnName)
         {
             if (dataTable.Rows.Count == 0)
                 return 0;
             var d = dataTable.Compute("Sum(" + columnName + ")", "");
             //return (decimal) d.ToString("C2");
-            return decimal.Parse(d.ToString());
+            return double.Parse(d.ToString());
         }
 
         public static List<T> ConvertDataTable<T>(DataTable dt) {
